updates
- delete ad api
- clean up auth.js and now:
when a user logs in or opens the app, it gets their location and
updates it.
- rid of phantom code in kijij.js
- in user.js:
added delete wishlist api. Not being called from anywhere yet
- in account.js:
added profile picture to schema and geo (for the user location)
- in categories.js
got rid of ads array. Never being used anywhere. And added a background
and icon field to the schema.
- kijiji.js (scraper)
fixed the issue where it wasn’t get the cars model field. And I got rid
of the block of code that pushes the ads id to the category’s ads array